Code space optimized
Min max door angles not determined

+++ Proto1/Project/ULP_SEC_CAM/Sensors/Include/lightSens_isl29035.h
@@ -107,7 +107,7 @@ uint16_t getLightsensor_data(void);
 //void configureISL29035(uint8_t ucAppMode);
 void configureISL29035(uint8_t ucAppMode,
  							uint16_t Lux_threshold, uint8_t Trigger);
-uint16_t verifyISL29035(void);
+int32_t verifyISL29035(void);
 uint16_t getLightsensor_intrptStatus(void);
 void setThreshold_lightsensor(uint16_t upper_threshold,
 								uint16_t lower_threshold);
